requirements.txt
asgiref==3.7.2
black==24.3.0
chardet==5.2.0
click==8.1.7
Django==5.0.3
fontawesomefree==6.5.1
mypy-extensions==1.0.0
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
reportlab==4.1.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
tzdata==2024.1
